third_party/WebKit/LayoutTests/fast/dom/custom/custom-element-name-sensivity.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(function () {
    var A = document.registerElement('mixed-CASE-custom-TAG');
    assert_true(document.createElement('MIXED-case-CUSTOM-tag') instanceof A);
    assert_true(
        document.createElement('MIXED-case-CUSTOM-tag', null) instanceof A);
}, 'createElement is case insensitive for custom tags');

test(function () {
    var B = document.registerElement('mixed-CASE-type-EXTENSION', {extends: 'span'});
    assert_true(document.createElement('span', 'MIXED-case-TYPE-extension')
                instanceof B);
}, 'createElement is case insensitive for type extensions');
</script>
